+++ isaaclab_arena_curobo/ik_reachability_validator.py
@@ -9,8 +9,7 @@ The pool's solve loop calls it on each geometry-valid candidate; a candidate is 
 top-down grasp at every movable object, so the loop keeps solving (reject-&-refill) until every env has enough reachable layouts.
 
 With the placement debug view on (``ObjectPlacerParams.debug_visualize``), the check also draws what it solved for each
-candidate -- the robot base, the grasps, their IK errors, and the collision spheres the arm holds at each solved grasp;
-see ``isaaclab_arena_curobo.reachability_visualizer``.
+candidate -- the robot base, the grasps, their IK errors; see ``isaaclab_arena_curobo.reachability_visualizer``.
 """
 
 from __future__ import annotations
